+++ services/director-v2/tests/unit/test_modules_dask_client.py
@@ -14,6 +14,7 @@ from uuid import uuid4
 
 import distributed
 import pytest
+import respx
 from _dask_helpers import DaskGatewayServer
 from _pytest.monkeypatch import MonkeyPatch
 from dask.distributed import get_worker
@@ -436,6 +437,7 @@ async def test_send_computation_task(
     image_params: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
     faker: Faker,
 ):
     _DASK_EVENT_NAME = faker.pystr()
@@ -520,6 +522,7 @@ async def test_computation_task_is_persisted_on_dask_scheduler(
     image_params: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
 ):
     """rationale:
     When a task is submitted to the dask backend, a dask future is returned.
@@ -605,6 +608,7 @@ async def test_abort_computation_tasks(
     image_params: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
     faker: Faker,
 ):
     _DASK_EVENT_NAME = faker.pystr()
@@ -682,6 +686,7 @@ async def test_failed_task_returns_exceptions(
     gpu_image: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
 ):
     # NOTE: this must be inlined so that the test works,
     # the dask-worker must be able to import the function
@@ -747,6 +752,7 @@ async def test_missing_resource_send_computation_task(
     image_params: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
 ):
 
     # remove the workers that can handle mpi
@@ -787,6 +793,7 @@ async def test_too_many_resources_send_computation_task(
     cluster_id: ClusterID,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
 ):
     # create an image that needs a huge amount of CPU
     image = Image(
@@ -825,6 +832,7 @@ async def test_disconnected_backend_raises_exception(
     cpu_image: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
 ):
     # DISCONNECT THE CLUSTER
     await dask_spec_local_cluster.close()  # type: ignore
@@ -854,6 +862,7 @@ async def test_changed_scheduler_raises_exception(
     cpu_image: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
 ):
     # change the scheduler (stop the current one and start another at the same address)
     scheduler_address = URL(dask_spec_local_cluster.scheduler_address)
@@ -892,6 +901,7 @@ async def test_get_tasks_status(
     cpu_image: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
     faker: Faker,
     fail_remote_fct: bool,
 ):
@@ -972,6 +982,7 @@ async def test_dask_sub_handlers(
     cpu_image: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
     fake_task_handlers: TaskHandlers,
 ):
     dask_client.register_handlers(fake_task_handlers)
@@ -1048,6 +1059,7 @@ async def test_get_cluster_details(
     image_params: ImageParams,
     mocked_node_ports: None,
     mocked_user_completed_cb: mock.AsyncMock,
+    mocked_storage_service_fcts: respx.MockRouter,
     faker: Faker,
 ):
     cluster_details = await dask_client.get_cluster_details()
